Migrate questions

+++ src/Quizer.Application/Quizes/Commands/DeleteQuestion/DeleteQuestionCommandHandler.cs
@@ -9,10 +9,12 @@ namespace Quizer.Application.Quizes.Commands.DeleteQuestion;
 public class DeleteQuestionCommandHandler : IRequestHandler<DeleteQuestionCommand, ErrorOr<QuestionId>>
 {
     private readonly IQuestionRepository _questionRepository;
+    private readonly IQuizRepository _quizRepository;
 
-    public DeleteQuestionCommandHandler(IQuestionRepository questionRepository)
+    public DeleteQuestionCommandHandler(IQuestionRepository questionRepository, IQuizRepository quizRepository)
     {
         _questionRepository = questionRepository;
+        _quizRepository = quizRepository;
     }
 
     public async Task<ErrorOr<QuestionId>> Handle(DeleteQuestionCommand request, CancellationToken cancellationToken)
@@ -20,6 +22,10 @@ public class DeleteQuestionCommandHandler : IRequestHandler<DeleteQuestionComman
         var question = await _questionRepository.Get(QuestionId.Create(request.QuestionId));
         if (question is null) return Errors.Question.NotFound;
 
+        var quiz = await _quizRepository.Get(question.QuizId);
+        if (quiz is null) return Errors.Quiz.NotFound;
+
+        quiz.DeleteQuestion(question);
         _questionRepository.Delete(question);
         question.Delete();
